requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
python-multipart>=0.0.6
requests>=2.31.0
twilio>=8.10.0
openai>=1.0.0
azure-cognitiveservices-speech>=1.34.0
aiofiles>=23.0.0
python-multipart>=0.0.6
faster-whisper>=1.0.0
